# Introduce an optimization in `CherryPick`