--- a/tla-bmcmt/src/main/scala/at/forsyte/apalache/tla/bmcmt/rules/aux/IntOracle.scala
+++ b/tla-bmcmt/src/main/scala/at/forsyte/apalache/tla/bmcmt/rules/aux/IntOracle.scala
@@ -21,33 +21,14 @@ import at.forsyte.apalache.tla.lir.values.TlaBool
  */
 class IntOracle(val intCell: ArenaCell, nvalues: Int) extends Oracle {
 
-  /**
-   * Produce an expression that states that the oracle values equals to the given integer position. The actual
-   * implementation may be different from an integer comparison.
-   *
-   * @param state
-   *   a symbolic state
-   * @param position
-   *   a position the oracle should be equal to
-   */
+  override def size: Int = nvalues
+
   override def whenEqualTo(state: SymbState, position: Int): TlaEx = {
     tla.eql(intCell.toNameEx, tla.int(position))
   }
 
-  /**
-   * Produce a ground expression that contains assertions for the possible oracle values.
-   *
-   * @param state
-   *   a symbolic state
-   * @param assertions
-   *   a sequence of assertions, one per oracle value
-   * @param elseAssertions
-   *   an optional sequence of assertions, one per oracle value
-   * @return
-   *   an expression ite(oracle = 0, ite(oracle = 1, ...))
-   */
   override def caseAssertions(state: SymbState, assertions: Seq[TlaEx], elseAssertions: Seq[TlaEx] = Seq()): TlaEx = {
-    if (elseAssertions.nonEmpty & assertions.size != elseAssertions.size) {
+    if (elseAssertions.nonEmpty && assertions.size != elseAssertions.size) {
       throw new IllegalStateException(s"Invalid call to Oracle, malformed elseAssertions")
     }
 
@@ -67,17 +48,6 @@ class IntOracle(val intCell: ArenaCell, nvalues: Int) extends Oracle {
     }
   }
 
-  /**
-   * Get a symbolic state and decode the value of the oracle variable into an integer. This method assumes that the
-   * solver context has produced an SMT model.
-   *
-   * @param solverContext
-   *   a solver context
-   * @param state
-   *   a symbolic state
-   * @return
-   *   an integer value of the oracle
-   */
   override def evalPosition(solverContext: SolverContext, state: SymbState): Int = {
     solverContext.evalGroundExpr(intCell.toNameEx).asInstanceOf[Int]
   }