--- a/tla-bmcmt/src/main/scala/at/forsyte/apalache/tla/bmcmt/rules/aux/IntOracle.scala
+++ b/tla-bmcmt/src/main/scala/at/forsyte/apalache/tla/bmcmt/rules/aux/IntOracle.scala
@@ -32,20 +32,7 @@ class IntOracle(val intCell: ArenaCell, nvalues: Int) extends Oracle {
       throw new IllegalStateException(s"Invalid call to Oracle, malformed elseAssertions")
     }
 
-    nvalues match {
-      case 0 => state.arena.cellTrue().toNameEx
-
-      case 1 => assertions.head
-
-      case _ =>
-        // iteCases is a sequence of tuples, with the fst and snd elements of each tuple being the "if" and "else" cases of an ite.
-        // If elseAssertions is not empty, each tuple has its fst element from assertions and its snd form elseAssertions.
-        // If elseAssertions is empty, each tuple has its fst element from assertions and its snd defaults to "ValEx(TlaBool(true))".
-        val iteCases = assertions.zipAll(elseAssertions, ValEx(TlaBool(true)), ValEx(TlaBool(true)))
-        val es =
-          iteCases.slice(0, nvalues).zipWithIndex.map { case (e, i) => tla.ite(whenEqualTo(state, i), e._1, e._2) }
-        tla.and(es: _*)
-    }
+    super.caseAssertions(state, assertions, elseAssertions)
   }
 
   override def evalPosition(solverContext: SolverContext, state: SymbState): Int = {