This PR refactors the `Oracle` interface, which we use to pick an arbitrary value from a sequence of elements or from a set. It also introduces a new oracle `ZipOracle` that compresses several values into a single one. By using `ZipOracle` in `CherryPick`, we can achieve a speed up on some benchmarks.

Below is a table for a small subset of benchmarks. We probably have to do large scale experiments later.


We have tried various configuration options:

 - `new`: using ZipOracle
 - `old`: not using ZipOracle
 - `bef`: using `--tuning-options=search.invariant.mode=before`
 - `aft`: using `--tuning-options=search.invariant.mode=after`
 - `nd`: using `--discard-disabled=0`

For some of the benchmarks, we ran 10-20 experiments with `hyperfine`, to
compute the average running times. The running times for such experiments look
like `t ± d`, that is, the expected running time lies in the time interval
`[t - d, t + d]`. These experiments demonstrate fluctuations of Z3.

## BMC up to 10 steps


|           | new/bef | old/bef | new/aft |old/aft | new/aft/nd | old/aft/nd |
| --------- | ------- | ------- | ------- | ------ | ---------- | ---------- |
| perf      |280 ± 25 |340 ± 104| 28 ± 4  | 30 ±12 |     14 ± 2 |    15 ± 3  |
| perf2     |9.35± 0.6|10.8± 1.2|8.2 ± 1.5|8 ± 1.2 |10.16 ± 4.7 |    9 ± 1   |
| EWD998-4  |183 ± 16 |   210   | 97 ± 7  |    106 | 95 ±  8    |     94     |
| ChangR-4  |  2952   |         |  2829   |        |      998   |     1105   |

## Inductive invariants

|           | new/bef | old/bef | new/aft |old/aft | new/aft/nd | old/aft/nd |
| --------- | ------- | ------- | ------- | ------ | ---------- | ---------- |
| EWD998-40 |154 ± 13 | 151 ± 7 |460 ± 66 | 478±75 | 441 ± 87   | 480 ± 45   |
| Bakery-3  | 32 ± 3  | 29 ± 5  | 20 ± 4  |19 ±1.6 | 17.5 ± 2   | 19.3 ± 3.4 |
 

- [x] Tests added for any new code
- [x] Ran `make fmt-fix` (or had formatting run automatically on all files edited)
- [ ] Documentation added for any new functionality
- [ ] Entry added to [UNRELEASED.md](./UNRELEASED.md) for any new functionality