@konnov 2022-05-10
> I found the move of comments to be exclusively on the parent class unexpected. Is this something we should aim to have in the whole codebase?

Yes, we had a discussion about this. I think it makes sense, as the comments of the overriding methods may get out of sync quick.